// packages/endpoints/src/server.options.d.ts
export interface ProxyOpts {
    proxyIpHeader: string;
    maxIpsCount?: number;
}
// ...
export interface ServiceConfig<TSerOpts = any> extends ServiceHandlerOptions<any>, TransferConfig {
    side: TransferSide.server;
    serverOpts?: TSerOpts;
    asDefault?: boolean;
    /**
     * is microservice or not.
     */
    microservice?: boolean;
    server?: any;
    /**
     * send detail error message to client or not.
     */
    detailError?: boolean;
    listenOpts?: any;
    proxy?: ProxyOpts;
    secure?: boolean;
    bootstrap?: boolean;
    payloadKey?: 'body' | 'payload';
    /**
     * configuration options.
     */
    config?: ConfigOptions;
    /**
     * configuration sources.
     */
    configSources?: ConfigSource[];
}